Move createRecord
